+++ db/migrations/000003_create_withdrawals_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE withdrawals
 (
     id           BIGSERIAL PRIMARY KEY    NOT NULL,
     order_number BIGINT                   NOT NULL,
-    sum          BIGINT                   NOT NULL,
+    sum          DOUBLE PRECISION         NOT NULL,
     user_id      BIGINT                   NOT NULL REFERENCES users (id) ON UPDATE CASCADE ON DELETE CASCADE,
     created_at   TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP NOT NULL,
     updated_at   TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP NOT NULL
