+++ internal/entity/order.go
@@ -13,11 +13,11 @@ const (
 )
 
 type Order struct {
-	CreatedAT time.Time     `db:"created_at"`
-	UpdatedAT time.Time     `db:"updated_at"`
-	Status    string        `db:"status"`
-	Number    int           `db:"number"`
-	Accrual   sql.NullInt64 `db:"accrual"`
-	ID        uint          `db:"id"`
-	UserID    uint          `db:"user_id"`
+	CreatedAT time.Time       `db:"created_at"`
+	UpdatedAT time.Time       `db:"updated_at"`
+	Status    string          `db:"status"`
+	Number    int             `db:"number"`
+	Accrual   sql.NullFloat64 `db:"accrual"`
+	ID        uint            `db:"id"`
+	UserID    uint            `db:"user_id"`
 }
